python-raspberry1/Gui1/requirements.txt
guizero~=1.2.0
Pillow~=7.0.0